+++ pajbot/modules/quest.py
@@ -43,6 +43,33 @@ class QuestModule(BaseModule):
                     'me',
                     'reply',
                     ]),
+            ModuleSetting(
+                key='reward_type',
+                label='Reward type',
+                type='options',
+                required=True,
+                default='tokens',
+                options=[
+                    'tokens',
+                    'points',
+                    ]),
+            ModuleSetting(
+                key='reward_amount',
+                label='Reward amount',
+                type='number',
+                required=True,
+                default=5,
+                ),
+            ModuleSetting(
+                key='max_tokens',
+                label='Max tokens',
+                type='number',
+                required=True,
+                default=15,
+                constraints={
+                    'min_value': 1,
+                    'max_value': 5000,
+                    }),
             ]
 
     def __init__(self):
@@ -92,7 +119,7 @@ class QuestModule(BaseModule):
         event = options['event']
         source = options['source']
 
-        message_tokens = '{0}, you have {1} tokens.'.format(source.username_raw, source.get_tokens())
+        message_tokens = '{0}, you have {1} tokens.'.format(source.username_raw, source.tokens)
 
         if self.settings['action_tokens'] == 'say':
             bot.say(message_tokens)
@@ -135,6 +162,7 @@ class QuestModule(BaseModule):
             return False
 
         self.current_quest = random.choice(available_quests)
+        self.current_quest.quest_module = self
         self.current_quest.start_quest()
 
         redis = RedisManager.get()
@@ -163,21 +191,6 @@ class QuestModule(BaseModule):
             log.error('No last stream ID found.')
             # No last stream ID found. why?
             return False
-
-        # XXX: Should we use a pipeline for any of this?
-        # Go through user tokens and remove any from more than 2 streams ago
-        for key in redis.keys('{streamer}:*:tokens'.format(streamer=StreamHelper.get_streamer())):
-            all_tokens = redis.hgetall(key)
-            for stream_id_str in all_tokens:
-                try:
-                    stream_id = int(stream_id_str)
-                except (TypeError, ValueError):
-                    log.error('Invalid stream id in tokens by {}'.format(key))
-                    continue
-
-                if last_stream_id - stream_id > 1:
-                    log.info('Removing tokens for stream {}'.format(stream_id))
-                    redis.hdel(key, stream_id)
 
     def on_loaded(self):
         if self.bot: