+++ pajbot/modules/quests/winhsbetwins.py
@@ -30,8 +30,6 @@ class WinHsBetWinsQuestModule(BaseQuest):
                     })
             ]
 
-    REWARD = 5
-
     def get_limit(self):
         return self.settings['quest_limit']
 
@@ -49,7 +47,7 @@ class WinHsBetWinsQuestModule(BaseQuest):
         redis = RedisManager.get()
 
         if user_progress == self.get_limit():
-            user.award_tokens(self.REWARD, redis=redis)
+            self.finish_quest(redis, user)
 
         self.set_user_progress(user.username, user_progress, redis=redis)